refactor(domain): CampusOpenTime 패키지 이동
