+++ src/main/java/domain/CampusOpenTime.java
@@ -1,4 +1,4 @@
-package constant;
+package domain;
 
 import java.time.LocalTime;
